requirements.txt
openai
python-docx